End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -17,7 +18,7 @@ export class MessagesComponent implements OnInit {
   loading = false
 
 
-  constructor(private messageService: MessageService) { 
+  constructor(private messageService: MessageService,private confirmService: ConfirmService) { 
 
   }
 
@@ -26,7 +27,6 @@ export class MessagesComponent implements OnInit {
   }
 
   loadMessages() {
-   
     this.messageService.getMessages(this.pageNumber, this.pageSize,this.container).subscribe(response => {
       this.loading = true;
       this.messages = response.result;
@@ -36,8 +36,12 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() =>{
-      this.messages.splice(this.messages.findIndex((m: any )=> m.id === id), 1);
+    this.confirmService.confirm('Confirm delete messages', 'This cannot be undone').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() =>{
+          this.messages.splice(this.messages.findIndex((m: any )=> m.id === id), 1);
+        })       
+      }
     })
   }
   pageChanged(event: any) {
